+++ pigeons/platform_api.dart
@@ -10,7 +10,8 @@ import 'package:pigeon/pigeon.dart';
     package: "com.adyen.checkout.flutter.generated",
     errorClassName: "AdyenPigeonError",
   ),
-  swiftOut: 'ios/adyen_checkout/Sources/adyen_checkout/generated/PlatformApi.swift',
+  swiftOut:
+      'ios/adyen_checkout/Sources/adyen_checkout/generated/PlatformApi.swift',
   swiftOptions: SwiftOptions(errorClassName: "AdyenPigeonError"),
   dartPackageName: 'adyen_checkout',
 ))
@@ -165,6 +166,14 @@ class AnalyticsOptionsDTO {
   );
 }
 
+class ThreeDS2ConfigurationDTO {
+  final String requestorAppURL;
+
+  ThreeDS2ConfigurationDTO(
+    this.requestorAppURL,
+  );
+}
+
 class DropInConfigurationDTO {
   final Environment environment;
   final String clientKey;
@@ -175,6 +184,8 @@ class DropInConfigurationDTO {
   final ApplePayConfigurationDTO? applePayConfigurationDTO;
   final GooglePayConfigurationDTO? googlePayConfigurationDTO;
   final CashAppPayConfigurationDTO? cashAppPayConfigurationDTO;
+  final TwintConfigurationDTO? twintConfigurationDTO;
+  final ThreeDS2ConfigurationDTO? threeDS2ConfigurationDTO;
   final AnalyticsOptionsDTO analyticsOptionsDTO;
   final bool showPreselectedStoredPaymentMethod;
   final bool skipListWhenSinglePaymentMethod;
@@ -193,6 +204,8 @@ class DropInConfigurationDTO {
     this.applePayConfigurationDTO,
     this.googlePayConfigurationDTO,
     this.cashAppPayConfigurationDTO,
+    this.twintConfigurationDTO,
+    this.threeDS2ConfigurationDTO,
     this.analyticsOptionsDTO,
     this.showPreselectedStoredPaymentMethod,
     this.skipListWhenSinglePaymentMethod,
@@ -399,6 +412,16 @@ class CashAppPayConfigurationDTO {
   );
 }
 
+class TwintConfigurationDTO {
+  final String iosCallbackAppScheme;
+  final bool showStorePaymentField;
+
+  TwintConfigurationDTO(
+    this.iosCallbackAppScheme,
+    this.showStorePaymentField,
+  );
+}
+
 class PaymentResultDTO {
   final PaymentResultEnum type;
   final String? reason;
@@ -508,6 +531,7 @@ class CardComponentConfigurationDTO {
   final AmountDTO? amount;
   final String? shopperLocale;
   final CardConfigurationDTO cardConfiguration;
+  final ThreeDS2ConfigurationDTO? threeDS2ConfigurationDTO;
   final AnalyticsOptionsDTO analyticsOptionsDTO;
 
   CardComponentConfigurationDTO(
@@ -517,6 +541,7 @@ class CardComponentConfigurationDTO {
     this.amount,
     this.shopperLocale,
     this.cardConfiguration,
+    this.threeDS2ConfigurationDTO,
     this.analyticsOptionsDTO,
   );
 }